+++ frontend/src/App.jsx
@@ -4,8 +4,8 @@ import { Route, Routes } from "react-router-dom";
 import Welcome from "./welcome";
 import Login from "./login";
 import Dashboard from "./dashboard";
-import AdminCreate from "./adminCreate";
-import UserCreate from "./userCreate";
+import AdminCreate from "./createQuestionnaire";
+import UserCreate from "./answerQuestionnaire";
 import Submitted from "./submitted";
 import "./styles/App.css";
 
@@ -18,6 +18,7 @@ export default function App() {
       <Route path="/create" element={<AdminCreate />} />
       <Route path="/newForm" element={<UserCreate />} />
       <Route path="/submitted" element={<Submitted />} />
+      <Route path="/createAdmin" element={<Submitted />} />
     </Routes>
   );
 }
